requirements.txt
aiohttp
aiosqlite
discord.py==2.5.2
python-dotenv
langchain
transformers
google-genai
langchain-google-genai
openai
requests
Pillow
chromadb